fix hoisting issue with total var
